python-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
opencv-python==4.8.1.78
mediapipe==0.10.8
trimesh==4.0.5
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
PyYAML==6.0.1
matplotlib==3.7.2
scipy==1.11.4
